Demo bars, formatted and colored
The demographics bars are about what I want. I don't know why the leading on the annotation gets smaller the further down the page you get, and I suspect that it is related to some error where the top and bottom bars in the figure are spacing out more. But I really need this figure to get into the report by the end of the day, and I suspect that I'm the only person reading this report who will be wringing their hands over typesetting issues.